+++ ee/tabby-ui/app/page/components/page-context.ts
