+++ ee/tabby-ui/app/page/components/section-title.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { HTMLAttributes, useContext } from 'react'
 
 import { cn } from '@/lib/utils'
